Remove "work arounds" from `dataAPIutilities.py` (#159)
